added descriptive ids to limiter functions

+++ src/utils.js
@@ -62,7 +62,7 @@ export async function fetchChannelMessages(channel, options, limiter) {
     while (!fetchComplete) {
         if (lastMessageId) fetchOptions.before = lastMessageId;
 
-        const fetched = await limiter.schedule(() => channel.messages.fetch(fetchOptions));
+        const fetched = await limiter.schedule({ id: "channel.messages.fetch" }, () => channel.messages.fetch(fetchOptions));
         if (fetched.size == 0) break;
 
         lastMessageId = fetched.last().id;
@@ -150,7 +150,7 @@ export async function fetchTextChannelData(channel, options, limiter) {
 
 /* creates a category for the guild */
 export async function loadCategory(categoryData, guild, limiter) {
-    const category = await limiter.schedule(() => guild.channels.create({ name: categoryData.name, type: ChannelType.GuildCategory }));
+    const category = await limiter.schedule({ id: "guild.channels.create" }, () => guild.channels.create({ name: categoryData.name, type: ChannelType.GuildCategory }));
     const finalPermissions = [];
 
     categoryData.permissions.forEach((permission) => {
@@ -164,7 +164,7 @@ export async function loadCategory(categoryData, guild, limiter) {
         }
     });
 
-    await limiter.schedule(() => category.permissionOverwrites.set(finalPermissions));
+    await limiter.schedule({ id: "category.permissionOverwrites.set" }, () => category.permissionOverwrites.set(finalPermissions));
     return category;
 }
 
@@ -172,7 +172,7 @@ export async function loadCategory(categoryData, guild, limiter) {
 export async function loadChannel(channelData, guild, category, options, limiter) {
 
     const loadMessages = async (channel, messages, previousWebhook) => {
-        const webhook = previousWebhook || await limiter.schedule(() => channel.createWebhook({ name: "MessagesBackup", avatar: channel.client.user.displayAvatarURL() }));
+        const webhook = previousWebhook || await limiter.schedule({ id: "channel.createWebhook" }, () => channel.createWebhook({ name: "MessagesBackup", avatar: channel.client.user.displayAvatarURL() }));
         if (!webhook) return;
 
         messages = messages.filter((message) => (message.content.length > 0 || message.embeds.length > 0 || message.files.length > 0)).reverse();
@@ -181,7 +181,7 @@ export async function loadChannel(channelData, guild, category, options, limiter
         for (let message of messages) {
             if (message.content.length > 2000) continue;
             try {
-                const sent = await limiter.schedule(() => webhook.send({
+                const sent = await limiter.schedule({ id: "webhook.send" }, () => webhook.send({
                     content: message.content.length ? message.content : undefined,
                     username: message.username,
                     avatarURL: message.avatar,
@@ -191,7 +191,7 @@ export async function loadChannel(channelData, guild, category, options, limiter
                     threadId: channel.isThread() ? channel.id : undefined
                 }));
 
-                if (message.pinned && sent) await limiter.schedule(() => sent.pin());
+                if (message.pinned && sent) await limiter.schedule({ id: "sent.pin" }, () => sent.pin());
             } catch (error) {
                 /* ignore errors where it request entity is too large */
                 if (error.message == "Request entity too large") return;
@@ -224,7 +224,7 @@ export async function loadChannel(channelData, guild, category, options, limiter
         createOptions.type = ChannelType.GuildVoice;
     }
 
-    const channel = await limiter.schedule(() => guild.channels.create(createOptions));
+    const channel = await limiter.schedule({ id: "guild.channels.create" }, () => guild.channels.create(createOptions));
     const finalPermissions = [];
 
     channelData.permissions.forEach((permission) => {
@@ -238,7 +238,7 @@ export async function loadChannel(channelData, guild, category, options, limiter
         }
     });
 
-    await limiter.schedule(() => channel.permissionOverwrites.set(finalPermissions));
+    await limiter.schedule({ id: "channel.permissionOverwrites.set" }, () => channel.permissionOverwrites.set(finalPermissions));
 
     if (channelData.type == ChannelType.GuildText) {
         let webhook;
@@ -249,7 +249,7 @@ export async function loadChannel(channelData, guild, category, options, limiter
 
         if (channelData.threads.length > 0) {
             channelData.threads.forEach(async (threadData) => {
-                const thread = await limiter.schedule(() => channel.threads.create({ name: threadData.name, autoArchiveDuration: threadData.autoArchiveDuration }));
+                const thread = await limiter.schedule({ id: "channel.threads.create" }, () => channel.threads.create({ name: threadData.name, autoArchiveDuration: threadData.autoArchiveDuration }));
                 if (webhook) await loadMessages(thread, threadData.messages, webhook);
             });
         }
@@ -261,39 +261,39 @@ export async function loadChannel(channelData, guild, category, options, limiter
 /* delete all roles, channels, emojis, etc of a guild */
 export async function clearGuild(guild, limiter) {
     const roles = guild.roles.cache.filter((role) => !role.managed && role.editable && role.id != guild.id);
-    roles.forEach(async (role) => await limiter.schedule(() => role.delete().catch((error) => console.error(`Error occurred while deleting roles: ${error.message}`))));
+    roles.forEach(async (role) => await limiter.schedule({ id: "role.delete" }, () => role.delete().catch((error) => console.error(`Error occurred while deleting roles: ${error.message}`))));
 
-    guild.channels.cache.forEach(async (channel) => await limiter.schedule(() => channel.delete().catch((error) => console.error(`Error occurred while deleting channels: ${error.message}`))));
-    guild.emojis.cache.forEach(async (emoji) => await limiter.schedule(() => emoji.delete().catch((error) => console.error(`Error occurred while deleting emojis: ${error.message}`))));
+    guild.channels.cache.forEach(async (channel) => await limiter.schedule({ id: "channel.delete" }, () => channel.delete().catch((error) => console.error(`Error occurred while deleting channels: ${error.message}`))));
+    guild.emojis.cache.forEach(async (emoji) => await limiter.schedule({ id: "emoji.delete" }, () => emoji.delete().catch((error) => console.error(`Error occurred while deleting emojis: ${error.message}`))));
 
-    const webhooks = await limiter.schedule(() => guild.fetchWebhooks());
-    webhooks.forEach(async (webhook) => await limiter.schedule(() => webhook.delete().catch((error) => console.error(`Error occurred while deleting webhooks: ${error.message}`))));
+    const webhooks = await limiter.schedule({ id: "guild.fetchWebhooks" }, () => guild.fetchWebhooks());
+    webhooks.forEach(async (webhook) => await limiter.schedule({ id: "webhook.delete" }, () => webhook.delete().catch((error) => console.error(`Error occurred while deleting webhooks: ${error.message}`))));
 
-    const bans = await limiter.schedule(() => guild.bans.fetch());
-    bans.forEach(async (ban) => await limiter.schedule(() => guild.members.unban(ban.user).catch((error) => console.error(`Error occurred while deleting bans: ${error.message}`))));
+    const bans = await limiter.schedule({ id: "guild.bans.fetch" }, () => guild.bans.fetch());
+    bans.forEach(async (ban) => await limiter.schedule({ id: "guild.members.unban" }, () => guild.members.unban(ban.user).catch((error) => console.error(`Error occurred while deleting bans: ${error.message}`))));
 
-    await limiter.schedule(() => guild.setAFKChannel(null));
-    await limiter.schedule(() => guild.setAFKTimeout(60 * 5));
-    await limiter.schedule(() => guild.setIcon(null));
-    await limiter.schedule(() => guild.setBanner(null));
-    await limiter.schedule(() => guild.setSplash(null));
-    await limiter.schedule(() => guild.setDefaultMessageNotifications(GuildDefaultMessageNotifications.OnlyMentions));
-    await limiter.schedule(() => guild.setWidgetSettings({ enabled: false, channel: null }));
+    await limiter.schedule({ id: "guild.setAFKChannel" }, () => guild.setAFKChannel(null));
+    await limiter.schedule({ id: "guild.setAFKTimeout" }, () => guild.setAFKTimeout(60 * 5));
+    await limiter.schedule({ id: "guild.setIcon" }, () => guild.setIcon(null));
+    await limiter.schedule({ id: "guild.setBanner" }, () => guild.setBanner(null));
+    await limiter.schedule({ id: "guild.setSplash" }, () => guild.setSplash(null));
+    await limiter.schedule({ id: "guild.setDefaultMessageNotifications" }, () => guild.setDefaultMessageNotifications(GuildDefaultMessageNotifications.OnlyMentions));
+    await limiter.schedule({ id: "guild.setWidgetSettings" }, () => guild.setWidgetSettings({ enabled: false, channel: null }));
 
     if (!guild.features.includes(GuildFeature.Community)) {
-        await limiter.schedule(() => guild.setExplicitContentFilter(GuildExplicitContentFilter.Disabled));
-        await limiter.schedule(() => guild.setVerificationLevel(GuildVerificationLevel.None));
+        await limiter.schedule({ id: "guild.setExplicitContentFilter" }, () => guild.setExplicitContentFilter(GuildExplicitContentFilter.Disabled));
+        await limiter.schedule({ id: "guild.setVerificationLevel" }, () => guild.setVerificationLevel(GuildVerificationLevel.None));
     }
 
-    await limiter.schedule(() => guild.setSystemChannel(null));
-    await limiter.schedule(() => guild.setSystemChannelFlags([
+    await limiter.schedule({ id: "guild.setSystemChannel" }, () => guild.setSystemChannel(null));
+    await limiter.schedule({ id: "guild.setSystemChannelFlags" }, () => guild.setSystemChannelFlags([
         GuildSystemChannelFlags.SuppressGuildReminderNotifications,
         GuildSystemChannelFlags.SuppressJoinNotifications,
         GuildSystemChannelFlags.SuppressPremiumSubscriptions
     ]));
 
-    await limiter.schedule(() => guild.setPremiumProgressBarEnabled(false));
+    await limiter.schedule({ id: "guild.setPremiumProgressBarEnabled" }, () => guild.setPremiumProgressBarEnabled(false));
 
-    const rules = await limiter.schedule(() => guild.autoModerationRules.fetch());
-    rules.forEach(async (rule) => await limiter.schedule(() => rule.delete().catch((error) => console.error(`Error occurred while deleting automod rules: ${error.message}`))));
+    const rules = await limiter.schedule({ id: "guild.autoModerationRules.fetch" },() => guild.autoModerationRules.fetch());
+    rules.forEach(async (rule) => await limiter.schedule({ id: "rule.delete" }, () => rule.delete().catch((error) => console.error(`Error occurred while deleting automod rules: ${error.message}`))));
 }
